♻️ 独立更新服务
