+++ lib/page/root/logic.dart
@@ -1,7 +1,7 @@
 import 'package:get/get.dart';
 
 import '../../manager/dun_preference.dart';
-import '../update/logic.dart';
+import '../../services/update_service.dart';
 import 'state.dart';
 
 class RootLogic extends GetxController {
@@ -17,7 +17,7 @@ class RootLogic extends GetxController {
     super.onInit();
     state.launchCount = (getLaunchCount() ?? 0) + 1;
     saveLaunchCount(state.launchCount);
-    UpdateLogic.to?.checkLatestVersion(autoCheck: true);
+    UpdateService.to?.checkLatestVersion(autoCheck: true);
   }
 
   @override
