+++ lib/page/update/state.dart
@@ -1,30 +1,23 @@
-import 'package:dio/dio.dart';
-
 import '../../common/data_status.dart';
 import '../../model/ceobe/version/dun_app.dart';
+import '../../services/update_service.dart';
 import '../../widget/progress.dart';
 
 class UpdateState {
-  String nowVersion = '0.0.0';
-  bool checking = false;
+  bool get isForce => UpdateService.to?.isForce ?? false;
+  String get nowVersion => UpdateService.to?.nowVersion ?? '';
+  String get newVersion => UpdateService.to?.dunAppInfo.version ?? '';
+  String get description => UpdateService.to?.dunAppInfo.description ?? '';
 
-  DunAppInfoModel dunAppInfo = DunAppInfoModel.fromJson({});
-  bool isForce = false;
-  String get newVersion => dunAppInfo.version;
-  String get description => dunAppInfo.description;
-
-  final List<DownloadUrlModel> downloadableUrlModels = [];
-  final List<DownloadUrlModel> manualUrlModels = [];
-  DownloadUrlModel? curDownloadUrlModel;
-
-  CancelToken? appCancelToken;
-  final ProgressController downloadProgressController = ProgressController();
-  DataStatus? downloadStatus;
-  String? downloadSavePath;
+  ProgressController? get downloadProgressController =>
+      UpdateService.to?.downloadProgressController;
+  DownloadUrlModel? get curDownloadUrlModel =>
+      UpdateService.to?.curDownloadUrlModel;
+  DataStatus? get downloadStatus => UpdateService.to?.downloadStatus.value;
+  List<DownloadUrlModel> get manualUrlModels =>
+      UpdateService.to?.manualUrlModels ?? [];
 
   String rootGID = "rootGID";
-  String downloadGID = "downloadGID";
-  String checkGID = "checkGID";
 
   UpdateState() {
     ///Initialize variables
